requirements.txt
spotipy
python-dotenv
matplotlib
